Move library persistence public methods to module
